bug(refs T32785): Move methods for pagination to mixin
To avoid code duplication these methods were moved to a mixin.
